+++ miniles-backend/pom.xml
@@ -6,7 +6,7 @@
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<version>2.5.3</version>
-		<relativePath/> <!-- lookup parent from repository -->
+		<relativePath/> <!-- lookup parent from CLDF_Est.minilesbackend.repository -->
 	</parent>
 	<groupId>CLDF_Est</groupId>
 	<artifactId>miniles-backend</artifactId>
