+++ miniles-backend/src/main/java/CLDF_Est/minilesbackend/service/erro/DocumentoNaoEncontrado.java
@@ -1,8 +1,8 @@
-package service.erro;
+package CLDF_Est.minilesbackend.service.erro;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
-public class TipoProposicaoNaoEncontrado extends RuntimeException {
+public class DocumentoNaoEncontrado extends  RuntimeException {
 }
