+++ miniles-backend/src/main/java/CLDF_Est/minilesbackend/service/erro/ProposicaoNaoEncontrada.java
@@ -1,4 +1,4 @@
-package service.erro;
+package CLDF_Est.minilesbackend.service.erro;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
