+++ miniles-backend/src/main/java/CLDF_Est/minilesbackend/service/mapper/DocumentoMappper.java
@@ -1,0 +1,9 @@
+package CLDF_Est.minilesbackend.service.mapper;
+
+import CLDF_Est.minilesbackend.domain.Documento;
+import org.mapstruct.Mapper;
+import CLDF_Est.minilesbackend.service.dto.DocumentoDTO;
+
+@Mapper(componentModel = "spring")
+public interface DocumentoMappper extends EntityMapper <DocumentoDTO, Documento> {
+}
