+++ miniles-backend/src/main/java/CLDF_Est/minilesbackend/service/mapper/ProposicaoMapper.java
@@ -1,8 +1,8 @@
-package service.mapper;
+package CLDF_Est.minilesbackend.service.mapper;
 
-import domain.Proposicao;
+import CLDF_Est.minilesbackend.domain.Proposicao;
 import org.mapstruct.Mapper;
-import service.dto.ProposicaoDTO;
+import CLDF_Est.minilesbackend.service.dto.ProposicaoDTO;
 
 @Mapper(componentModel = "spring", uses = {TipoProposicaoMapper.class, DocumentoMappper.class} )
 public interface ProposicaoMapper extends EntityMapper<ProposicaoDTO, Proposicao> {
