--- miniles-backend/src/main/java/service/dto/TipoProposicaoDTO.java
@@ -1,20 +1,0 @@
-package service.dto;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class TipoProposicaoDTO {
-
-    private Long id;
-    private String nome;
-    private String sigla;
-
-
-}
