--- miniles-backend/src/main/java/service/mapper/DocumentoMappper.java
@@ -1,9 +1,0 @@
-package service.mapper;
-
-import domain.Documento;
-import org.mapstruct.Mapper;
-import service.dto.DocumentoDTO;
-
-@Mapper(componentModel = "spring")
-public interface DocumentoMappper extends EntityMapper <DocumentoDTO, Documento> {
-}
