--- miniles-backend/src/main/java/service/mapper/TipoProposicaoMapper.java
@@ -1,9 +1,0 @@
-package service.mapper;
-
-import domain.TipoProposicao;
-import org.mapstruct.Mapper;
-import service.dto.TipoProposicaoDTO;
-
-@Mapper(componentModel = "spring")
-public interface TipoProposicaoMapper extends EntityMapper<TipoProposicaoDTO,TipoProposicao> {
-}
