+++ miniles-backend/src/main/resources/application.yml
@@ -6,4 +6,4 @@ spring:
     password: min
   liquibase:
     change-log: classpath:/config/liquibase/master.xml
-#    drop-first: true
+    drop-first: true
