+++ miniles-backend/src/main/resources/config/liquibase/changelog/20210817133400_documento.xml
@@ -7,8 +7,8 @@
             </column>
             <column name="nome" type="varchar(40)"/>
             <column name="texto" type="varchar(300)"/>
-
             <column name="id_proposicao" type="bigint"/>
+
 
         </createTable>
     </changeSet>
