+++ miniles-frontend/src/app/components/proposicao/proposicao.component.scss
@@ -1,5 +1,1 @@
-ng-template tr{
-    color: blue;
 
-    background-color: black;
-}
